Basic Login Functionality Completed
Impana worked on the login prompt itself, but had issues getting the data out of the window

I implemented the IPC communication, which DOES need to be changed as it has some bad security implications at the moment.

But we have the ability to send info from the authentication window that heimdall uses to our backend.

Co-Authored-By: impanadev <51289348+impanadev@users.noreply.github.com>