Functional Tide Prompt for Tide Datasource Logins
Some issues and concerns:
If you close the application then open it again (meaning you don't have a valid jwt as you've yet to log in) the stored tide vault still unlocks for some reason.
You can now add multiple vaults if you use multiple different tide accounts (this works just fine SO far, I suspect this is due to the fact that the encrypt decrypt is static at the moment, although we do store the jwt in the source data... so this may be the reason for it working as well?)
I've implemented a crude reauth requirement if the token is invalid and I ensure it's invalid when you launch buttercup.
I've done some testing in regards to the different accounts and 'reauthenticating' as an incorrect account (they still decrypt, which makes sense due to static decryption) but they upload the saved data correctly, and I suspect that's due to the fact that we store vaults in the database using the UUID, which is static and buttercup is able to remember which UUID belongs to where using temp data of some sort (something I'm not 100% on)